change vec3 vec4 constructor from {} to () !!! add obj parser, ignore texture and material (ka kd ks kr = 0 for now), parser untested pass compilation, orignal features working

+++ as2/camera.cpp
@@ -3,8 +3,8 @@
 camera::camera(vec3 eye, vec3 ll, vec3 lr, vec3 ul, vec3 ur) 
   :eye(eye), ll(ll), lr(lr), ul(ul), ur(ur) {
     double ar = (lr-ll).norm()/(ul-ll).norm();
-    screen_w = (ar > 1)? 800*ar:800;
-    screen_h = (ar > 1)? 800:800*ar;
+    screen_w = static_cast<unsigned int>(ar > 1 ? 800 * ar : 800);
+    screen_h = static_cast<unsigned int>((ar > 1)? 800:800*ar);
   }
 
 ray camera::generate_ray(double u, double v) const {
